requirements.txt
python==3.8
pandas==1.0
biopython==1.78
lxml==4.5
